Hustle Full Project
An on/off endeavor to learn Godot by making a clone of Panel de Pon / Tetris Attack. Development began around September of 2018 and became 'functionally complete' by the end of March 2019 judging by the file histories. The basic game flow allows two local players to compete in endless runs with rematch or quit-to-menu options. Settings can be changed to increase the speed of the game grid rising, how many blocks are available, and the window scale.